+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxBody.hs
@@ -18,7 +18,7 @@
 module Cardano.Ledger.Shelley.TxBody (
   ShelleyDelegCert,
   Delegation (..),
-  ConstitutionalDelegCert (..),
+  GenesisDelegCert (..),
   MIRCert (..),
   MIRPot (..),
   MIRTarget (..),
@@ -106,6 +106,7 @@ import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.PParams (Update)
 import Cardano.Ledger.Shelley.TxCert (
+  GenesisDelegCert (..),
   MIRCert (..),
   MIRPot (..),
   MIRTarget (..),